+++ samples/client/petstore/python/src/petstore_api/components/schema/format_test.py
@@ -93,14 +93,14 @@ class ArrayWithUniqueItemsTuple(
 ArrayWithUniqueItemsTupleInput = typing.Union[
     typing.List[
         typing.Union[
-            float,
-            int
+            int,
+            float
         ],
     ],
     typing.Tuple[
         typing.Union[
-            float,
-            int
+            int,
+            float
         ],
         ...
     ]
